requirements.txt
streamlit
torch
Pillow
transformers
